Abstracted the Escrow tests to a behaviour
