+++ test/payment/Escrow.test.js
@@ -1,73 +1,11 @@
-const BigNumber = web3.BigNumber;
+import shouldBehaveLikeEscrow from './Escrow.behaviour';
 
-require('chai')
-  .use(require('chai-bignumber')(BigNumber))
-  .should();
-
-const EVMRevert = require('../helpers/EVMRevert.js');
 const Escrow = artifacts.require('Escrow');
 
-contract('Escrow', function ([payer1, payer2, payee1, payee2]) {
-  const amount = web3.toWei(42.0, 'ether');
-
+contract('Escrow', function (accounts) {
   beforeEach(async function () {
     this.contract = await Escrow.new();
   });
 
-  it('can accept a single deposit', async function () {
-    await this.contract.deposit(payee1, { from: payer1, value: amount });
-
-    const balance = await web3.eth.getBalance(this.contract.address);
-    const deposit = await this.contract.deposits(payee1);
-
-    balance.should.be.bignumber.equal(amount);
-    deposit.should.be.bignumber.equal(amount);
-  });
-
-  it('reverts on empty deposits', async function () {
-    await this.contract.deposit(payee1, { from: payer1, value: 0 }).should.be.rejectedWith(EVMRevert);
-  });
-
-  it('can add multiple deposits on a single account', async function () {
-    await this.contract.deposit(payee1, { from: payer1, value: amount });
-    await this.contract.deposit(payee1, { from: payer2, value: amount * 2 });
-
-    const balance = await web3.eth.getBalance(this.contract.address);
-    const deposit = await this.contract.deposits(payee1);
-
-    balance.should.be.bignumber.equal(amount * 3);
-    deposit.should.be.bignumber.equal(amount * 3);
-  });
-
-  it('can track deposits to multiple accounts', async function () {
-    await this.contract.deposit(payee1, { from: payer1, value: amount });
-    await this.contract.deposit(payee2, { from: payer1, value: amount * 2 });
-
-    const balance = await web3.eth.getBalance(this.contract.address);
-    const depositPayee1 = await this.contract.deposits(payee1);
-    const depositPayee2 = await this.contract.deposits(payee2);
-
-    balance.should.be.bignumber.equal(amount * 3);
-    depositPayee1.should.be.bignumber.equal(amount);
-    depositPayee2.should.be.bignumber.equal(amount * 2);
-  });
-
-  it('can withdraw payments from any account', async function () {
-    let payeeInitialBalance = await web3.eth.getBalance(payee1);
-
-    await this.contract.deposit(payee1, { from: payer1, value: amount });
-    await this.contract.withdraw(payee1, { from: payer2 });
-
-    const escrowBalance = await web3.eth.getBalance(this.contract.address);
-    const finalDeposit = await this.contract.deposits(payee1);
-    const payeeFinalBalance = await web3.eth.getBalance(payee1);
-
-    escrowBalance.should.be.bignumber.equal(0);
-    finalDeposit.should.be.bignumber.equal(0);
-    payeeFinalBalance.sub(payeeInitialBalance).should.be.bignumber.equal(amount);
-  });
-
-  it('reverts on empty withdrawals', async function () {
-    await this.contract.withdraw(payee1, { from: payer1 }).should.be.rejectedWith(EVMRevert);
-  });
+  shouldBehaveLikeEscrow(accounts);
 });
